Update RNIapIos.m
There was no purchase failed callback being called. Added one that returns a JS object with a numeric error code and an English description. This is important because it allows developers the ability to not only handle when the transaction is not complete, but also handle it appropriately. For instance, when error.code == 2, it means the user canceled it. You don't need to alert them that they canceled it. But if error.code == 4, you may want to tell them that their device is currently set to not allow IAPs. Maybe they are children. Who knows. And of course, restore any locked app functionality.

Co-authored-by: foxvector <bivector16@gmail.com>
Co-authored-by: checkmate17 <hubqueen410@gmail.com>